(bug 1073264) allow the use of X-Sendfile for serving attachments
see bug 906010, bug 1073241 and bug 1073246.
This change allows an administrator to configure bugzilla to offload
serving attachments directly to nginx/apache/lighttpd.
This is done simply by returning a special header with a filepath.

This change gives a know in the admin/params/attachments to select
with header to use. The code in attachment.cgi checks whether the
said attachment is on filesystem and if both conditions are met,
outputs the header rather than the file content.

nginx -> X-Accel-Redirect,
apache2/httpd -> X-Sendfile,
lighttpd -> X-LIGHTTPD-send-file.

I've personally tested with apache2 + mod_xsendfile,
as well as nginx + X-Accel-Redirect.