Bug 1283612 - Add Bugzilla::Extension module loader last
Currently, anything that is loaded after the bugzilla extensions
are loaded has to search through each callback inserted into @INC
by Bugzilla::Extension. If you have a lot of extensions this is slow.

It is harmless to search for Bugzilla::Extension::* modules last, so
we replace the unshift with a push.