<option> used when no flag is set is displayed as 'X' in Safari
After https://github.com/bugzilla/bugzilla/commit/7d0c10e71498f2909f466fc0ccda1890595bd3ff,
the <option> used when no flag is set is displayed as 'X' in Safari. This is because WebKit
displays the label when provided as per the latest HTML specification:
- https://html.spec.whatwg.org/multipage/rendering.html#the-select-element-2
- https://html.spec.whatwg.org/multipage/forms.html#concept-option-label

Chrome/Blink also display the label in standards mode, but not in quirks mode. The proposed
fix is to use a whitespace as label instead of 'X'. This renders as empty in Firefox / Chrome
and Safari and is still recognized as valid HTML as per https://validator.w3.org.