replace perltidyrc with on that specifies -pbp with two exceptions
previously, the perltidyrc was written with the original recommendations of the
upstream bugzilla project. The style used there is... inconsistent at best. For
a while now, I've been using perltidy -pbp and telling others to do the same for
new code. However there are two exceptions to -pbp that are widely useful:

1. line length can be up to 120.
2. brace vertical tigntness is turned off.
   This means that there is always a newline after the { in a multi-line brace
   construct.