Bug 1447027 - Change default mail method for vagrant and update README
Updates the default mail_delivery_method to be Sendmail for vagrant
development. This allows developers to connect to the imap server
running at bmo-web.vm:143 to view mail using a real mail client.

The default method remains the same for docker users, 'Test'.

Both methods are described in the README in a new section.