Bug 1478983 - WebService endpoint to check if a user can enter a bug into a given product
There is some common code, which we will need again for another
WebService call, that can be refactored into their own functions.

This WebService endpoint takes a product name and user ID in the form
`check_enter_bug/<product>/<user id>`.  It returns an object with
one attribute, `result`, that is set to 1 if the product exists and
the user is allowed to enter bugs in it, or 0 otherwise.