no bug - use more generous timeouts
this makes all the mojo timeouts larger -- except for 'clients' which should be smaller because we're so synchronous. It also puts them into environmental variables so ops can tweak them.

Note some of the code has moved to the main application class to make future people less likely to not notice these values.