Bug 1546444 - Parameterize time that the security bugs report uses
Prior to this commit the report could specify a date to base the report
of off, but, it didn't accept a time (hour, min, second).

With this commit the report can be based on the exact day and time it
was run instead of just 0000 UTC of the specified date.

E.g. secbugsreport.pl $(date +'%Y %m %d %H %M %S %z').