[Bug 1592129] Pass a reference to db into Schema object
This change adds a 'db' attribute to the `Bugzilla::DB::Schema` class.
In two places (`get_empty_schema` and `deserialize_abstract`) the invocant
argument was called `$class` but was never actually a class. Those were renamed
to `$self`. It was fortunate that they were always objects because otherwise it
would have been more difficult to ensure a `db` is always present.