ensure that mysql keywords are always quoted
This patch uses `Bugzilla::DB::QuoteIdentifier` and a new
`Bugzilla::DB->quote_columns` method to ensure we always quote column names when
they would be reserved words in MySQL 8 and later.

This does not mean we quote every bare word, but in any situation where the
value might be a keyword (e.g. **groups**) it is quoted.

The goal here is to not make a larger diff than required.