allow preview of comment in new bugs
FreeBSD Bug 250699 is titled " Preview tab is blank for new issues" (q.v.)

This suggested workaround seems to fix this problem in practice.

Note: the changed code is only applicable to 5.0.4 and 5.2 -- the file is completely rewritten in harmony.  For this reason, using this as a workaround, rather than finding the underlying bug it obviates, is probably sufficient for existing Bugzill installations.

tbd: evaluate whether this fixes Bugzilla Bug 1247165.